requirements.txt
Django>=2.2,<2.3
wagtail>=2.7,<2.8
django-crispy-forms==1.8.1
django-allauth==0.40.0
wagtailmenus==3.0
django-modelcluster==5.0
django-taggit==1.1.0